[native] Delete files not needed for the debug build
The debug build for PrestoC++ fails as we are running out of disk.
The disk space is limited and this is one way of reducing the size.
Tests are not being built here and can be removed.

In addition, a newer, smaller dependency image is being used to reduce
the disk space consumption even more.