[native] Switch to using the Velox mono library
This change makes use of the Velox mono library for non-test related
components. This removes the need to add individual velox
subcomponent libraries and cuts down on the linking size because symbols
are duplicated.

When adding new non-test velox dependencies use the velox library rather
than the individual velox component.
For tests, the velox components are still needed to be separately used.