feat: Add fast path for single bit in ByteOutputStream::appendBits (#12047)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/12047

While working on speeding up the PrestoBatchVectorSerializer I noticed that sometimes we write a single bit to
ByteOutputStreams (e.g. when writing a single null or the null flag).  ByteOutputStream::appendBool has a fast path for this
case, which ByteOutputStream::appendBits is missing, this provides a minor speed up that adds up over time.

Reviewed By: spershin

Differential Revision: D67988365

fbshipit-source-id: 96f67e2ad18e3de54e1a279385d636b825c734ba