refactor: Break PrestoSerializer.cpp into components (#11922)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/11922

PrestoSerializer.cpp has grown into a roughly 5000 line behemoth. It makes navigating the file a challenge.

This change is a pure refactor and does not introduce any changes besides moving things between files and renaming functions. It breaks the file up the code into 8 new files based on function:
* PrestoIterativeVectorSerializer: contains the PrestoIterativeVectorSerializer class
* PrestoBatchVectorSerializer: contains the PrestoBatchVectorSerializer class
* PrestoHeader: contains the PrestoHeader class which is used for reading/parsing the header of a PrestoPage
during deserialization
* VectorStream: contains the VectorStream class which is an appendable container for serialized values used
during serialization
* PrestoVectorLexer: contains the PrestoVectorLexer class along with Token and TokenType, this is used to lex
serialized PrestoPages for the benefit of compression libraries like Zstrong.
* PrestoSerializerDeserializationUtils: contains the logic needed to deserialize a PrestoPage, notably the
function readTopColumns which serves as the entry to deserialization
* PrestoSerializerEstimationUtils: contains the logic needed to estimate the size of a serialized PrestoPage, this
logic is shared by PrestoBatchVectorSerializer and PrestoVectorSerde's estimateSerializedSize function (used
with PrestoIterativeVectorSerializer)
* PrestoSerializerSerializationUtils: contains the generic logic needed to serialize a PrestoPage, notably the
function serializeColumn which serves as the entry point to serialization and the function flushStreams used
write the final output to an OutputStream.  This logic is shared by PrestoBatchVectorSerializer,
PrestoIterativeVectorSerializer, and PrestoVectorSerde's deserializeSingleColumn function.

These are all internal details of the PrestoSerializer (which is why they were in the cpp file) so I've placed all the
moved logic in a "detail" namespace (with the exception of Token and TokenType in PrestoVectorLexer).

I've verified that all tests pass and run the benchmark to verify I didn't introduce any regressions.

Reviewed By: xiaoxmeng, spershin

Differential Revision: D67543330

fbshipit-source-id: b2b7017a991b0f7abf7afd35459c7537e128dc27