misc(fuzzer): Allow avg(interval) in aggregation fuzzer (#12077)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/12077

The avg() function uses `row(double, bigint)` as the intermediate states, so the ordering of
inputs could cause a small precision error to the double field in the intermediate states and
hence the final result. https://github.com/facebookincubator/velox/pull/12048 adds a new signature to allow `avg(interval) -> row(double, bigint) -> interval`.
However, because the `interval` type is an integral type, the imprecision in the aggregation
result is not tolerated by the assertEqualResults() API, hence causing fuzzer failures, such
as https://github.com/facebookincubator/velox/actions/runs/12698298641/job/35397706898.
This diff adds a custom result verifier for avg() that cast the interval results to double before
calling assertEqualResults(), so that the epsilon-comparison instead of regular comparison is
used in assertEqualResults().

Reviewed By: kgpai

Differential Revision: D68117666

fbshipit-source-id: bbdc4b38b37e59f79e9cd10d6eeca93221f2e054