fix: Set_agg should not throw on nested nulls (#12093)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/12093

Velox's implementation of set_agg throws when it detects nulls nested inside complex types,
while Presto Java tolerates them.

set_agg was changed to throw on nested nulls as part of this issue which discovered that Velox
and Presto Java had different semantics when it came to nested nulls in a variety of
aggregations.
https://github.com/facebookincubator/velox/issues/6314

After this was done, Presto Java changed the semantics of some aggregations to be like "distinct
from" rather than equals for a handful of aggregations/UDFs (i.e. treat null as a value and don't
throw).
https://github.com/prestodb/presto/issues/22900

set_agg was one of these, and it doesn't look like anyone went back and undid the change in
Velox so we allow nested nulls.

We should check the rest of the functions mentioned in that issue, but I think most weren't
implemented in Velox when the Presto Java change was made, so they were likely done correctly
from the beginning.

Reviewed By: kgpai

Differential Revision: D68226492

fbshipit-source-id: 74b0db975a4a1f6f5f84b95688ce29e12d0d93c7