test: Reenable set_agg and set_union in AggregationFuzzer (#12098)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/12098

set_agg and set_union were disabled in AggregationFuzzer because Velox and Presto 0.288 had
different ordering semantics.  This was fixed in 0.289 and we bumped the version of Presto we
compare against to 0.290 so this is no longer a concern.

Ran a biased AggregationFuzzer run against only set_agg and set_union comparing against
Presto for 1 hour and didn't see issues.

Reviewed By: spershin

Differential Revision: D68286084

fbshipit-source-id: 3a49438467a8fdd64b0a5ef9828058a5ba6f9100