test: Add a case with LocalPartition using hash partitioning in JoinFuzzer (#12112)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/12112

Today JoinFuzzer only tests plans with LocalPartition that uses round robin partitioning. This
sends the entire Vector to a single thread in addInput, which means we aren't testing the case
where LocalPartition wraps the input in a Dictionary.

I discovered this while investigating the issue fixed in
https://github.com/facebookincubator/velox/pull/12090  This case, combined with ASAN, is
enough to catch that issue. (I was able to demonstrate this by applying this change to the
commit just before the fix).

Reviewed By: bikramSingh91

Differential Revision: D68342589

fbshipit-source-id: f918db30b32a63c5ef917f14ea9d930c1b7380f4