fix(merge-join): Properly load lazies on right join (#12141)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/12141

When executing a right outer merge join, we cannot produce a buffer
containing LazyVectors if the current left buffer may still participate in
subsequent records. This can cause an issue where the lazy vector is
materialized by an operator downstream only for a few records, and the merge
join code may still try to reference records that were not materialized,
leaving garbage in the output.

To fix it, ensure that whenever we detect we crossed the boundary to the next
right batch and an output batch is produced, that left lazies are loaded.

Reviewed By: mbasmanova

Differential Revision: D68480018

fbshipit-source-id: 3aada0b7d8dbb9f07bb5706565fa958231e94c49