fix(fuzzer): Remove hardcoded table names in Fuzzers (#12120)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/12120

Looking into the failed fuzzer job in D68359107 shows that the table name, "t_values", in the query does not match the table name of the created table, "tmp". This is because of a change made in D66977480 which updated the toSql(ValuesNode) method to return "t_<id>" instead of "tmp". That change ensured unique table names since PlanNode ids are unique within a plan.

The change in this diff deprecates hardcoding "tmp" as table names in place of using toSql. A toSql method was also added for TableScanNodes. There are many instances of this hardcoding which makes this diff a bit large.

Reviewed By: spershin, kagamiori

Differential Revision: D68400743

fbshipit-source-id: 142610df19875c1cdba16921cdf4f816c4ac8aa5