fix: regexp_extract returns match in mismatched group (#12109)
Summary:
The implementation of Re2Extract has a bug, that it might consider a mismatched group as MATCHED empty string "" rather than MISMATCHED std::nullopt.

For example, in the function calling: regexp_extract("rat cat\nbat dog", "ra(.)|blah(.)(.)", 2).
In this case, for group 2 the result must be std::nullopt because no substring would match pattern `blah(.)`.
But the current implementation would mistake the matching of group 1 `ra(.)` as a empty match case for group 2, and thus return a empty matching, which is wrong.

See detailed description in https://github.com/facebookincubator/velox/issues/12119.

This PR fix this bug in Re2Extract implementation.

Also note that this bug behavior exists in Re2ExtractAll as well, but this PR doesn't modify in Re2ExtractAll because existing UTs of Re2ExtractAll already rely on this behavior.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/12109

Reviewed By: kgpai

Differential Revision: D68334118

Pulled By: kagamiori

fbshipit-source-id: 5824102903760c0459af64a6fe62dd195c5f1928