[native] Suppress connector values from being logged
The log contains the connector key and value pairs read from the
config files. The problem is that certain key-value pairs likely contain
secrets that should not be logged as this leaks credentials.

We cannot filter the keys to be able to print non-secret information
because there is no consistency for naming schemes of these properties
and we could miss newly ones added.