feat: Add new FileFormat::SST to toFileFormat and toString (#12137)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/12137

A new file format FileFormat::SST was introduced, but it was not added to these methods (toFileFormat and toString). Adding a new case statement to the toFileFormat function to return the FileFormat::SST value when the input string is "sst". The toString function is also updated to include a new case statement to return "sst" when the file format is FileFormat::SST. This will allow users to specify the file format as "sst" when working with Velox.

Reviewed By: kKPulla, sdruzkin, harsharastogi

Differential Revision: D68463647

fbshipit-source-id: e8af42237fa4fd278bbe46a28d1dfee2aefbaacc