feat: Avoid small batches in Exchange (#12010)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/12010

Prevent exchange client from unblocking to early. Unblocking to early impedes
effectiveness of page merging. When the cost of creating a vector is high (for
example for data sets with high number of columns) creating small pages can
make queries significantly less efficient.

For example it was observed that when network is congested and Exchange buffers
are not filled up as fast query may experience CPU efficiency drop up to 4x: T211034421

Reviewed By: xiaoxmeng

Differential Revision: D67615570

fbshipit-source-id: a9936d9a95d43045f3bdabc8db5a4638bfef177c