misc(fuzzer): Support custom result verifier with compare() API in window fuzzer (#12148)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/12148

The custom result verifiers of some functions provides a compare() API that can be used to
compare the expected and actual results after certain transformation. This is needed for
functions like `avg(interval) -> interval` because the result of this function can have a small
precision error that should be tolerated. However, the window fuzzer used to only support
the verify() API of custom result verifiers. This diff makes the window fuzzer to use the
compare() API as well when available.

Reviewed By: kgpai

Differential Revision: D68507422

fbshipit-source-id: cae34c2936335e5639e5ec7af00fa80c36c39bd6