[native] Fix varchar cast for json (#24396)
Summary:

https://www.internalfb.com/tasks/?t=211442303

There was an error in running query on Prestissimo not Presto - "Scalar function presto.default.substr not registered with arguments: (JSON, BIGINT, BIGINT)".

This is not due to missing function, as the function signature does not exist in Presto. It occurs when attempting to cast JSON as varchar of capped length.
Related Diff: https://www.internalfb.com/diff/D59531026
Note: Exception is still raised for try_cast() behavior. Alignment is out of scope for this PR

Differential Revision: D68353517