build(cmake): Make monolithic libs be built inside Velox's own binary directory when being used as a subproject (#12144)
Summary:
Similar to https://github.com/facebookincubator/velox/pull/12128, currently when Velox is imported as a subproject of another project, the monolithic libs `libvelox.a` / `libvelox.so` will be generated in root project's binary directory which is sub-optimal.

The patch fixes the issue so the libs will be generated inside Velox's binary directory.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/12144

Reviewed By: kgpai

Differential Revision: D68606631

Pulled By: xiaoxmeng

fbshipit-source-id: d0ef042c5ee935a598637b07549c9c48a06708fe