fix: Fix withIntDistributionForField copy (#11644)
Summary:
For decimal(8, 5) data build, Arrow reports below error due to out of range.
The reason is in the copy of 'withIntDistributionForField', the target index
and source index are reversed.
```
`C++ exception with description "Invalid: Invalid cast from Decimal128 to 4 byte integer" thrown in the test body.
```

Pull Request resolved: https://github.com/facebookincubator/velox/pull/11644

Reviewed By: Yuhta

Differential Revision: D66560309

Pulled By: bikramSingh91

fbshipit-source-id: bd594f67a58b92915ea3bf24bee3a311db11eb04