refactor: use tryFindOrCompile instead of findOrCompile for regex [1/n] (#12234)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/12234

When evaluating regexes and the regex throws, we can simply capture the exception inside a status instead of doing a throw which is extremely expensive. This change just uses the pre-existing expected API for error handling. This allows queries which times out after 2 hours on an operator to finish in 47 minutes.

Reviewed By: Yuhta

Differential Revision: D68983392

fbshipit-source-id: d54a7f76a6b33df05f6561092cc5c942c00a409c