fix(parquet): Clear column chunk metadata of rowgroups not selected (#10225)
Summary:
Parquet file metadata contains the metadata of all the rowgroups in the file. However, the processing
happens per split which likely consists of a few rowgroups. This memory is not tracked and can result in OOMs.
The scope of this PR is to release the columns chunks (have the most memory footprint) of RowGroups that
are not selected.
TPC-DS results show good memory utilization due to this change and prevented OOMs.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/10225

Reviewed By: DanielHunte

Differential Revision: D68979191

Pulled By: bikramSingh91

fbshipit-source-id: 6b41961a177c9b5805d64df5ab45cd389d8b9119