feat: Add DecodedVector::sharedBase() (#12249)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/12249

Sometimes we need to take shared ownership of the base value vector of a dictionary.  The current `DecodedVector` only keeps reference to a raw pointer so there is no way to get hold of the `shared_ptr`.  We add `DecodedVector::decodeAndGet()` so that we can get the shared ownership.

Reviewed By: bikramSingh91

Differential Revision: D69081492

fbshipit-source-id: 94fc4afb3995c0fe80c7ae77909e4fde144e2680