fix: Address ipprefix type incompatibility mismatch (#12145)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/12145

Java represents ipprefix as 17 byte varchar with the first 16 bytes as BE form ipaddress, and the 17th byte as the prefix length.

In Velox, we represent ipprefix as a struct/row of <IPADRESS, TINYTINY> which is internally represented as a ROW<HUGEINT, TINYINT>. Natively, this is represented as int128_t for the ipaddress in LE form and int8_t for the prefix.

The mismatch in types causes serialization and deserialization issues because the encoding for java is VARIABLE_WIDTH encoding but Velox is sending ROW encoding.

To address this, let's transform the velox representation of IPPREFIX to match that of Java. In other words, transform ROW<IPADDRESS, TINYINT> -> VARIABLE_WIDTH (16 Byte BE form of ip, 1 byte tiny).

Reviewed By: Yuhta

Differential Revision: D68285595

fbshipit-source-id: dc441ba978ff6f3dd1c5d8f127d4d090c6f2eaa1