fix: Ensure lambda functions fail on oversized array inputs (#12271)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/12271

We recently encountered an issue where a lambda function (filter)
having an array as a capture column resulted in creating an oversized
array vector. When inside the lambda eval, the capture is replicated by
wrapping it in a dictionary that aligns each row with each element
of the array that the filter is applied to. It then attempted to
flatten this wrapped vector, where it first calculated the size of the
elements buffer to create which overflowed and resulted in creation
of a buffer much smaller than the required size. Finally when filling
in the values in this elements buffer it ended up writing outside the
bounds of the buffer and caused a seg fault.

This is a bandaid fix for this issue to ensure we fail instead of
crashing.

Reviewed By: kevinwilfong

Differential Revision: D69210640

fbshipit-source-id: 47f4e6844a78717d4c04bad181471239855cc718