[native] Add clang installation to dependency image and update image
PR #24479 switched the debug CI build to Clang to save more disk space.
Clang is installed as part of the CI steps but it is better to have it present in the
dependency image. The other change was made to unblock PRs due to failing CI pipelines.

In addition, the dependency image is updated with the change made in this PR.