fix: Fixe handling of user errors in json_parse (#12272)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/12272

json_parse currently only properly propagates user errors that come from
normalizedSizeForJsonParse or from simd when parsing the JSON.

User errors can come from elsewhere in the UDF, e.g. the generateViews logic, which are not
caught and thus fail the query even when the UDF is wrapped in a "try".

To fix this I've wrapped the bulk of the logic that gets executed on a single row in a try catch that
handles user errors, propagating them to the EvalCtx.

Reviewed By: Yuhta, kgpai

Differential Revision: D69211907

fbshipit-source-id: 2ee7422c0f80a35b6f44221fe4942ab2cc04ae3a