feat: Add setting/getting properties on directories. (#12247)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/12247

We'd like to set/get properties on directories if the remote filesystem supports APIs like TTL. In POSIX, this would be similar to ioctl.

Simply expose a property bag for which users can use for setting key-value pairs for the set path, and on the get path, only support getting a single key for now. If the key is not present, nullopt is returned.

Reviewed By: singcha, zation99

Differential Revision: D68650034

fbshipit-source-id: 761318b2068734bc8a7c871ff9da626537bc2e1d