[iceberg] Reformat incorrectly formatted files
As a part of the project, we have a set of checkstyle guidelines.
Generally we recommend developers to use IntelliJ for Java
development. If you use the official style guidelines and format
with the provided style rules within IntelliJ there are a number of
files in the Iceberg connector which don't follow the official
formatting rules. Maven checkstyle isn't capable of finding them
all automatically.

The changes in this commit are the result of running IntelliJ's reformat
procedure on all the files in the Iceberg connector using Presto's
officially-provided style rules.