fix: wildcard support in json extract using simdjson (#12281)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/12281

This change addresses 2 bugs in the wildcard support of json_extract:

Bug 1: Failure to iterate on values inside an object, resulting in
empty arrays.
For eg: calling JSON_EXTRACT(
   JSON '{"foo": {"a" : {"c": 1 , "d": 3} }, "b": 2}', '$.foo.a.[*]')
will return an empty array whereas we expect [1, 3]

Bug 2: Inconsistent handling of indefinite results, always returning a
JSON array even when the token is not evaluated.
or eg: JSON_EXTRACT('({"a": {"b": [123, 456]}})', '$.a.c[*]') should
return NULL instead of [] as the key 'c' does not exist.

Reviewed By: kevinwilfong

Differential Revision: D69281233

fbshipit-source-id: 4e4e604a6d22a9e81785fb2a998f7508a770a736