fix: UserErrors leave json_parse in an invalid state (#12308)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/12308

When json_parse throws a user error and it's not wrapped in a try, if that exception is caught
outside expression evaluation and the ExprSet attempts to be reused, the json_parse function
may be in an invalid state and will throw a runtime exception.

This was exposed in ExpressionFuzzer after
https://github.com/facebookincubator/velox/pull/12281 was landed.

This fix is to clear the state whenever we see a UserException after calling parse whether or not
the function is wrapped in a try.

Reviewed By: kgpai

Differential Revision: D69473523

fbshipit-source-id: d4ae34d65136ce80138f9ed2e7f6f86bf566011e