fix(s3): Allow custom cacheKey function (#12217)
Summary:
Today, S3FileSystem caches the instances based on the endpoint value.
The endpoint value is not a sufficient key in practice. Usually, the credentials are
per bucket. Presto uses the bucket name as a key as well.
This PR changes the default key to use endpoint + bucket name.
Users can optionally provide a function to define the key
for caching the instances to allow customization.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/12217

Reviewed By: Yuhta

Differential Revision: D69406902

Pulled By: kgpai

fbshipit-source-id: 332cf807124acf691196b2407a6610a37a743ffd