fix[RowContainer]: Fix bug in RowContainer::extractValuesWithNulls (#12301)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/12301

Bug happens when:
We have a target (result) vector which has some rows and not a single null
(null buffer is not allocated).
We extract to that target vector a column from a RowContainer and that column
has null in at least one row.

Bug manifests in all rows that the target vector had before extracting our
RowContainer into it becoming nulls.

This is because the null buffer was allocated with 'setNotNull' set to false.

We fix it by supplying true when 'resultOffset' is positive (meaning that
target vector already has some rows).

Reviewed By: Yuhta

Differential Revision: D69445179

fbshipit-source-id: 457f050d8e6773691f0682195b30114a2328de01