fix: Fix operator finishTiming metrics (#12316)
Summary:
Issue: We record the execution time of `nextOp->noMoreInput()` to `op`'s
finishTiming metrics. This issue is noticeable in the OrderBy, Window, and
HashBuild operators, where computationally intensive tasks are executed within
the `noMoreInput()` function—for example, sorting occurs in `noMoreInput()`. As
a result, the recorded wall time is incorrectly added to upstream operators.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/12316

Reviewed By: kKPulla

Differential Revision: D69635556

Pulled By: pedroerp

fbshipit-source-id: a31c41976f196c007de2140b790c590f359c8bda