fix: Fix optimization of re-using inputVector (#12332)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/12332

D69567930 introduced an optimization to reuse the inputVector to save cycles and likely memory when extracting data from a group.

However, there is a subtle bug here when we call `extractSingleGroup()`, it can return 0 rows. In this case, we must still update the firstInputColumn, which is a rolling index that represents which inputVectors map to which aggregateInputs.

By not updating this, we may get a mismatch in the reused inputVector and the actual underlying type at this particular column. This is a hidden bug that surfaces if the following conditions are at least satisfied:

1. We've extracted a group with 0 rows
2. We've retried to use an index at a position which has a type that is different that the column index.

Reviewed By: Yuhta

Differential Revision: D69625562

fbshipit-source-id: d252510c03687867fd9a443e8dc52500bfd01936