feat: Add option to HiveDataSink to always create a file (#12331)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/12331

This change adds an option to HiveDataSink (plumbed via HiveInsertTableHandle) to always
create a file.  This is useful for query engines that write bucketed tables by running a Task per
output bucket rather than by setting the HiveBucketProperty. In this case, even if there is no data
to process, Velox will still produce an empty file for the output bucket.

Note that this option is not possible to support when the HiveBucketProperty is set or there a
partition keys in the data, as this means that the number of files/directories respectively is
determined at runtime from the input data. Checks for this are enforced at plan generation time.

Reviewed By: Yuhta

Differential Revision: D69629165

fbshipit-source-id: d3ea37c4ad58c1562e3940f35595ffdfade70c90