misc(fuzzer): Log fuzzer stats to stderr (#12328)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/12328

Fuzzer jobs in GitHub Actions currently only print logs to stderr to screen and save logs to
stdout to the artifact. Because the fuzzer stats provide useful signals about the test
coverage, we'd like to print them on the screen so that developers do not have to download
artifacts to see them.

Reviewed By: kgpai

Differential Revision: D69624523

fbshipit-source-id: dc2e9d3e15018d53338917eed1a4bb5fbc8ed1e0