fix: Reduce lock cost in AsciiInfo updates (#12363)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/12363

In some cases (e.g. row-wise deserialization) we need to resize the
string vector frequently.  In that case none of the ASCII info is valid, but we
still pay the price to lock and check the state.  Fix the cost by using an
atomic boolean for fast path check.

Reviewed By: kevinwilfong

Differential Revision: D69750090

fbshipit-source-id: c930fb10ed0c2a38fcbd1bb9a55bde39fe65cbfe