fix: Fix grouping key reordering during spilling (#12395)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/12395

When prefix sort is enabled, we sort the grouping keys to maximize the prefixsort benefit as introduced in https://github.com/facebookincubator/velox/pull/11720.

However, when spilling happens and when reading the spilled data, there are key order mismatch between the spilled data and the operator output. It can cause segmentation fault when there is RowType mismatch or other type mismatch failures.

This PR fixes by adding a spill data loader (`spillResultWitoutAggregates_`) which has the reordered grouping keys, loading the spilled data, and mapping the keys back to result after loading.

Reviewed By: xiaoxmeng

Differential Revision: D69860326

fbshipit-source-id: ca1aa35cfe08124709a72cf4eeddb0b2b3731102