feat(fuzzer): Allow functions to be tested only with sorted input in aggregation fuzzer (#12392)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/12392

The tdigest_agg() function needs to be tested only with input of determined order so that Velox result is comparable with Presto's. This is because different ordering of input values affects the content in TDigest. Therefore, this diff makes it possible to specify a set of functions to be tested only with ordered inputs in aggregation fuzzer. Aggregation fuzzer will then use `order by arg0, arg1, ...` inside the aggregaiton function calls of this function.

Reviewed By: natashasehgal

Differential Revision: D69886108

fbshipit-source-id: 304d13ae41cbc14df330a02cc2a775c731929637