test(fuzzer): Skip map_keys_by_top_n_values Test (#12447)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/12447

The fuzzer test for map_keys_by_top_n_values should be skipped until the presto array comparison null catch patch is fully released to the public.

The only failure scenario occurs when the input value is an array with a null element, which is low risk and has already been fixed in this  https://github.com/prestodb/presto/pull/24570. Once the patch is fully rolled out, we can remove the method from the skip list.

Reviewed By: kgpai, yuandagits, HeidiHan0000

Differential Revision: D70198160

fbshipit-source-id: 834457779074ab255ecbd28b77427a4ff75b7b74