fix(tpch): Properly generate rows for lineitem (#12427)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/12427

When using dbgen, the number of lineitems generated for each order is
chosen at random. To be able to control rows ids and generate data in parallel
given the definition on splits, lineitem data is generated based on the number
of associated orders, so it needs special handling logic on the reader. Adding
the required logic so that lineitem data can be generated in parallel as
expected.

Reviewed By: xiaoxmeng

Differential Revision: D70029685

fbshipit-source-id: f7ac8bc3e1c39ef1905afedf4181089c36882389