Fix DateTimeFormatter to set weekDateFormat only if it has week and day of week along with year. (#12456)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/12456

This is a targeted fix that ensures that date_parse formats which contain %a but without additional specifiers like %v does not set weekDateFormat. An example of such a format is  '%d-%b-%Y (%a)' which when parsing a string like '11-Mar-2024 (Mon)' will incorrectly ignore the month and day and instead return a date like '2024-01-01' .

There are likely other cases where additional or inconsistent specifiers are handled incorrectly. A full fix of DateTimeFormatter is beyond scope of this PR, but I will send out additional PR's in the future to address these issues.

Mitigates S493660.

Reviewed By: pedroerp

Differential Revision: D70253624

fbshipit-source-id: 0462cd525edede0d32c4b8942e60b11b8dfc7223