feat(fuzzer): Add TopNRowNumberFuzzer (#12103)
Summary:
topNRowNumber node is an optimized planNode for SQL with ranking window functions but which limits them to only the topN results. Add a TopNRowNumberFuzzer for plans with this planNode.

This fuzzer is closely modeled after the RowNumberFuzzer. So the common code is abstracted to a RowNumberFuzzerBase class which is used as the parent class for both RowNumberFuzzer and TopnRowNumberFuzzer.

The fuzzer generates plans only for row_number function right now. It will be enhanced to support rank and dense_rank functions after https://github.com/facebookincubator/velox/pull/11554

Pull Request resolved: https://github.com/facebookincubator/velox/pull/12103

Reviewed By: xiaoxmeng

Differential Revision: D69936162

Pulled By: kagamiori

fbshipit-source-id: 81214748c874f219cf7bc57b5eeeb8039325b06c