misc: Enforce the max size of a write to spill file (#12446)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/12446

In PrestoSQL,  spilling can expose degenerate structures which are over 2GB (int32_t) in size. While being over int32_t does not have intrinsic issues, the Presto wired format requires that the size of the payload be at most 4 bytes. This means that for PrestoSQL case, we cannot spill if the payload is > int32_t. Moreover, many of the serializers have a limit of int32_t. For now, on the write path, just fail the write if a single write is > int32_t.

Reviewed By: xiaoxmeng

Differential Revision: D70187465

fbshipit-source-id: 2f8517bcf23ae6eba328ed068f81e1cba85192a6