fix(sparksql): Respect session timezone when casting varchar to timestamp (#11975)
Summary:
When casting varchar to timestamp, the timezone embedded in input or the
oneconfigured by user should be respected in the conversion. The proposed
changesfor SparkSQL are quite similar to that for Presto, except the parse
mode. So in thispr, a common function was also extracted for doing timezone-
based adjustment.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/11975

Reviewed By: kgpai

Differential Revision: D70581205

Pulled By: kevinwilfong

fbshipit-source-id: 32e80b2e8fb2fdfa7c261aa488d45ece218d81d1