Add ClassLoaderSafeConnectorPlanOptimizer
This is required when plugin optimizer rules have a dependency which
may rely on a properly set ContextClassloader in order to perform
any of the optimizations.

In the case of Iceberg, there are cases when we can reach out for some
of the table's metadata files which initiates calls to the FileSystem
client. The Hadoop FileSystem Configuration object uses the
ContextClassLoader to load the correct FileSystem client implementation.
We found in production when many connector are enabled and the query
spans multiple connectors, there are cases where the ContextClassLoader
is not set properly resulting in ClassNotFound exceptions