fix: Presto UDF split_part does not match Presto behavior for empty string delimiter (#12583)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/12583

This fixes a bunch of behavior inconsistencies between Presto Java's and Velox's split_part UDFs.

In Presto Java, when split_part is called on a string X with an empty string delimiter and index i, it
returns the i'th character of X or null if X has fewer than i characters. It also validates that every
character up to and including the i'th character is valid UTF-8 (or all characters if it returns null).

In Velox today, it will return the entire string X if the i is 1 and the empty string otherwise.

In Velox today, we do not check that the index is a positive value, unlike Presto Java.

In Velox today, we do not check that the search string contains valid UTF-8 characters. Note that
Presto does not check this condition when the delimiter is non-empty, Velox maintains this
behavior after this change.

Reviewed By: yuandagits

Differential Revision: D70810224

fbshipit-source-id: c55c7dbb3eb05fc3091a2690a997f78371b9d751