test: The new tzdb library can handle DST into the distant future [3/4] (#12471)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/12471

One of the goals of adopting time zone library from LLVM's std::chrono implementation was to
support DST into the distant future. There are a bunch of tests that check that we enforce a limit
on datetimes in time zones that have DST. This PR removes those tests (that were commented out
in the original PR) and replaces them with tests that ensure we do in fact support these datetimes
and handle DST correctly.

Reviewed By: yuandagits

Differential Revision: D70337030

fbshipit-source-id: d2ea1a9a62788e0f7f123dda0764d71c6ec4e76a