Remove native Presto-on-Spark
The main motivation for this is to support the upgrade of Presto
from JDK 8 to 17. Native Presto-on-Spark native relies on airlift's
http client which will be upgraded to require Java 17 support in
the near future, making Presto-on-Spark incompatible with Java 8.

This is a no-go as certain users require Spark to remain on Java 8

Since native PoS is effectively unused, this is the solution to
move forward. In the future, when users are ready to upgrade PoS
to Java 17 support, we could revert this PR to re-introduce native
PoS support.