Fix path filtering for matrix jobs
Due to new github branch protection rules, both Java 8 and 17 matrix
jobs have to pass. Our old path filter rules prevent the matrix jobs
from running if the path filter doesn't pass (e.g. doc-only PRs). In
order to resolve this, we add a check at the beginning of each step
for the path filter. It is not ideal, but github does not provide a
better way at the moment