fix: The new LLVM library for time zones is slow with DST in the distant future (#12636)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/12636

The new LLVM library has logic to handle rules that apply forever (this was something missing
from the old date library that drove the need to pull in this new library).

Unfortunately this logic is very slow with years far in the future, this is because of the way it looks
for the rule that applies at a given time.  It starts from the beginning of the continuation and
calculates the next time a transition happened between rules, checks if the new rule applies to
the target time and so on. It needs to do this because the start time of a given rule is calculated
using the previous rule, so it needs to find the start time of the previous rule, and so on.

With these forever rules there's only ever at most 2 that apply and they alternate on an annual
basis. We can take advantage of this to speed things up.

We look at the previous year before the target date. We figure out which one has an earlier
transition time (we know their transition times are calculated using the other rule since they
alternate). Then we can start from the earlier of the two meaning we only have to iterate through
the loop a few times, instead of 10'000s of times.

This may not be the most efficient solution, but it's good enough, and feels safer than anything
more revolutionary.

Reviewed By: yuandagits

Differential Revision: D71140180

fbshipit-source-id: 07ed3bcb2fd63642365ac4a2348aa8a5065905dc