refactor: Clean up some error messages and file structure (#12657)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/12657

In addition to some text changes and extra invariant checks, BingTileType will soon
get a bunch more logic.  So I've extracted the non-inline `bingTileInvalidReason`
from BingTileType.h to BingTileType.cpp.  This is the only type in prestosql that
has a cpp file, so I'd like to check that's the right direction to go.

Reviewed By: bikramSingh91

Differential Revision: D71222628

fbshipit-source-id: 68e94bdd7d85965256bec7aa691a75ee76bb36e3