Remove descriptor mapping from table function analysis and planning
This is a change of approach on resolving arguments
of table functions.
Before this change, the mapping of descriptor arguments
to table arguments had to be determined during query
analysis. It was too limiting: in that model, input table
references were limited to the specified descriptor fields.

After this change, any input fields can be referenced,
and the required fields (channels) are to be determined
dynamically during function compilation.

One advantage of determining the required fields early
was the possibility to prune any columns that were
not required by the table function. This should be replaced
by adding a way for a table function to declare required
fields during query optimization.

Changes adapted from trino/PR#13602
Original commit: cbde4d7168efddf1d56f906d3d1e50d542d2f851
Author: kasiafi

Co-authored-by: kasiafi <30203062+kasiafi@users.noreply.github.com>