refactor: Move json_extract_scalar back to simple function (#12608)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/12608

We initially moved json_extract_scalar to a vector function
implementation so that we can reuse the JSON parsing logic and apply
it to VARCHAR inputs. JSON parsing achieves the following for VARCHAR
input: (1) it formats spaces, (2) it sorts the keys inside objects,
(3) it flags invalid JSON, and (4) it unescapes string scalars. For
json_extract_scalar, we only care about (3) and (4) since the output
is a scalar. Therefore, we can avoid the costly operation of
rewriting the large input JSON in a normalized form and only apply
the operations we require. To achieve (3), SIMD JSON can be used to
quickly check for valid JSON, and for (4), SIMD JSON provides an API
to extract a string scalar after removing escaping.

This change moves json_extract_scalar back to simple function
implementation and ensures (3) and (4) are applied.

Reviewed By: kgpai

Differential Revision: D70978059

fbshipit-source-id: 1fe7246121896b25d9a6f6206c954c2a6847535e