fix: MergeJoin fails when adding nulls for a miss to an input column that is constant (#12713)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/12713

In MergeJoin's prepareOutput function we call wrapInDictionary on any input projections so
we can add duplicates and nulls. However, if the input column is a ConstantVector this will
simply return a resized ConstantVector rather than a DictionaryVector.

This becomes a problem when we want to add a null for a miss, we can't call setNull on a
ConstantVector.

To fix this I've added a check, when we want to add a null, if the Vector is a ConstantVector
we force it to be wrapped in a DictionaryVector (by allocating a nulls buffer which we're going
to need to add the null to anyway).

The thinking is that in general misses should be rare, and there's a small performance benefit
to using a ConstantVector rather than wrapping it in a DictionaryVector unless we have to.

Reviewed By: kKPulla

Differential Revision: D71438385

fbshipit-source-id: 2cd0661173d6fc0fad61e7c7e7ab8c2bfeaad4e1