fix(build): Arrow bundled dependency build may fail on macOS (#12658)
Summary:
An issue occurs if thrift is installed to the system. On macOS we don’t install arrow and it is always bundled. If thrift was installed in the system (e.g. via brew) PR 12599 changed the flow such that FindThrift is called twice. thrift::thrift is first added via FindThrift in FindArrow line 20. Then again FindThrift is called when the Cmakefile for Arrow is called line 13. This is removed because we now always call FindThrift. Prior to PR 12599 it would not have called FindThrift if Arrow was not present.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/12658

Reviewed By: kKPulla

Differential Revision: D71410174

Pulled By: kagamiori

fbshipit-source-id: f694cb53f4fae8f08eb8b47df6f0b334f347af5a