Remove airlift:http-client from presto-main-base
This requires a refactor of the airlift:discovery and
airlift:event dependencies and their users from
presto-main-base into presto-main

The main issue is that the ExchangeClient requires both
drift and HTTP client implementations. We can't just move the
ExchangeClient class because lots of code would move with it.

To address this, the ExchangeClient has been modified to
instead accept an RpcShuffleClientFactory object which
is capable of producing a shuffle client implementation
for the ExchangeClient. This factory is plumbed through via
guice. This allows us to move the client implementation into
presto-main, but leave the ExchangeClient itself inside of
presto-main-base. This is OK because Presto-on-Spark
does not use the ExchangeClient.