Prevent extra license checks on presto-root
This adds a configuration for the license-maven-plugin to skip
checks against unnecessary files in presto-root, but still enables
it for all child modules.

The main issue is that when enabled for presto-root, the plugin will
check *all* child directories, even if it is already covered by the
plugin execution at a later stage. When running `./mvnw validate`
the build spends a significant chunk of time checking all files under
presto-root, then continues on to do  the validations for all other
modules.

On my MacBook Pro M1 processor I found the following performance
improvements after experimenting with 5-6 builds:

Before:
- ./mvnw validate: 44-48 seconds (average ~45 seconds)
After:
- ./mvnw validate 30-33 seconds (average ~31 seconds)