fix: Ensure proper lifetime of stats in DirectCoalescedLoad (#12756)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/12756

There is a crash in HiveDataSource during the merging of stats because the stats may still be used by the executor in DirectBufferedInput. This means we need the stats to have lifetime exceed that of the executor inside DirectBufferedInput. The simple fix here is to pass a shared_ptr of the stats to DirectBufferedInput so that DirectCoalescedLoad owns a reference to the stats.

DwioCoalescedLoadBase was updated to also cache the stats shared_pointer since it can go on the same code path.

Reviewed By: Yuhta

Differential Revision: D71641158

fbshipit-source-id: cd114c020c88eac3b8c9921b9df41c55f4bdaa28