feat(s3): Add log location configuration (#12534)
Summary:
The S3 filesystem uses the default logging configuration which creates the file in the current directory where an executable is called from. Most deployments have a custom location to write logs.
This change allows the user to specify a path that is used as the new log prefix which results in an absolute path being used for the log file creation. The default AWS log file prefix is used.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/12534

Reviewed By: zation99

Differential Revision: D71669762

Pulled By: xiaoxmeng

fbshipit-source-id: 727623cf071b10b548f87c48100cad36bfe4aa0b