Support ExchangeNode within DeleteNode
Summary:
Prior to this change, processing of DeleteNode in TableWriteInfo was assumed to have a direct descendant of TableScanNode, with both included in the same plan fragment.  However, this is not the case if a partitioned exchange is involved in the DELETE plan.  In that case the DeleteNode and TableScanNode can be present in different fragments.

After this change, if the TableScanNode is present in the same fragment, it is returned.  Otherwise, we stop traversing children at RemoteSourceNode and return an empty Optional for TableScanNode.  As a result, the DeleteScanInfo in TableWriteInfo will also be set to an empty Optional.

Differential Revision: D71692096