fix: Expand the range of supported years (#12694)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/12694

Presto Java supports any timestamp that can fit in the 52 bits well allocate for millis UTC in
the TimestampWithTimeZone type.  Likewise the Timestamp type can support dates with
years past 200 million.

In Velox currently we limit the year to what can be represented with 16 bits (this is what
std::chrono applies in C++20).

This change adds support for all years where all dates in the year can be expressed as a 64
bit signed integer millis since epoch. (Note that this is 1 year less than Presto Java on either
end, but Presto Java silently overflows).

Reviewed By: yuandagits

Differential Revision: D71349259

fbshipit-source-id: 08765337cd8e76576f440cb9ca68dda086ab5f4a