fix(fuzzer): Make less ambiguous function not execution by adding to unary operator map (#12786)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/12786

Update query generation for function `not` to treat it as an operator and not a function. Currently it is being used as the following which is leading to errors:
```
SELECT (BOOLEAN 'false' >= not(BOOLEAN 'true'));

Presto query failed: 1 line 1:28: mismatched input 'not'.
```
We can wrap the function to "unambiguize" the call:
```
presto:tpch> SELECT (BOOLEAN 'false' >= (not BOOLEAN 'true'));
 _col0
-------
 true
(1 row)
```

Reviewed By: kagamiori

Differential Revision: D71767079

fbshipit-source-id: 8501fd01a25a02cdaf5f3476c3819ef7094e509a