fix: PrestoSerde should be using chain length during serialization (#12784)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/12784

Compression does not guarantee a single buffer. This caused a silent corruption (see in test plan) because we were passing in the length of the single iobuf length for the compressed data instead of the whole data chain size. This caused issues on the decompression path with fail to decompress or worse, silent errors (but I havn't seen this).

Fix is to simply pass the whole chain length.

Reviewed By: xiaoxmeng, singcha

Differential Revision: D71761561

fbshipit-source-id: 59ca0b41b9ebcd8318be0a14a15762dfe5fa5404