Fix flaky test: TestBrutalShutdown.testQueryRetryOnShutdown
When debugging this tests' failures, I found that some queries were
failing with an error that is not marked as retriable, causing this test
to fail. By changing the respective error code to one that is retriable
this test no longer seems to fail.

This may have occurred due to changes in the way the task event
loops are scheduled, but I am not 100% sure. Either way, I think this
error code is capable of query retries, so the solution seemed suitable.