fix: Prevent null-deref in complex vector during estimateFlatSize and prepareForReuse (#12767)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/12767

size_ and offset_ are set to nullptr in prepareForReuse if they are not mutable; however, size_ and offset_ should always be valid.

The typical usage pattern of `prepareForReuse()` is that caller wants to save memory and resource, they call `prepareForReuse()` and then `resize()`; however, if caller does not call `resize()` and call functions which depend on `size_` and `offset_`, they can crash.

This diff fixes up `prepareForReuse()` by creating an buffer for `size_` and `offset_`  with BaseVector::length_ if the underlying buffer is not mutable instead of setting it to null.

Reviewed By: xiaoxmeng, Yuhta

Differential Revision: D71689808

fbshipit-source-id: ca0a410aee29fce44201e8c797c73e6b9b8418ad