feat(orc): Adding ORC functionality into StripeStream Info (#12738)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/12738

Logarithm is intending to migrate to Velox and we need to support our existing ORC format.  Long term we can investigate migrating to DWRF formate however to unblock the near term efforts we need to have backward compatibility with our existing ORC format.

For this change set we are updating Velox to support reading ORC format

Core Changes
- Added StripeFooterWrapper to handle proto::StriperFooter and proto::orc::StripeFooter
- Added added StreamKind to indepentantly identify proto::Stream_Kind vs proto::orc::Stream_Kind
- Added overloads and helper to correctly get the DwrfStreamIdentifier depending dwrf vs orc EncoderKey
- Updated all readers to use the correct EncoderKey to get StreamId (dwrf vs orc)

# diff
This diff adds some of the core functionality for exposing ORC stripefooter to the stream
- Adds column encoder to the DwrfStreamIdentifier
- Parses orc::StripeFooter to a Stream

Next diff we will handle the column reader functionality and enable the orc::StripeFooter to be parsed into the StripeFooterWrapper

Reviewed By: Yuhta

Differential Revision: D71331177

fbshipit-source-id: 686a5abadc018ad95f71ae2c6200b6d56b1b2042