feat(orc): Adding missing support for reading orc files (#12739)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/12739

Logarithm is intending to migrate to Velox and we need to support our existing ORC format.  Long term we can investigate migrating to DWRF formate however to unblock the near term efforts we need to have backward compatibility with our existing ORC format.

For this change set we are updating Velox to support reading ORC format

Core Changes
- Added StripeFooterWrapper to handle proto::StriperFooter and proto::orc::StripeFooter
- Added added StreamKind to indepentantly identify proto::Stream_Kind vs proto::orc::Stream_Kind
- Added overloads and helper to correctly get the DwrfStreamIdentifier depending dwrf vs orc EncoderKey
- Updated all readers to use the correct EncoderKey to get StreamId (dwrf vs orc)

# diff
This should be a no functional changes for reading DWRF format.

This diff add support for exposing ORC encodings to all of the reader functionality and enables the orc::StripeFooter to be parsed into a StripeFooterWrapper

Reviewed By: Yuhta

Differential Revision: D71048852

fbshipit-source-id: 7aeaee5a7084ab2615356b6b357329963676a871