fix(fuzzer): Wrap is null call to remove query ambiguity (#12820)
Summary:
A previous change updated `is_null` SQL converstion to `is null` since the prior is not registered with Presto. This solved the issue, but also created a rare case where the SQL executor may be confused by the order of operations (see Test Plan). This change will wrap this conversion in parantheses to avoid any ambiguity.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/12820

Reviewed By: kKPulla

Differential Revision: D71906852

Pulled By: peterenescu

fbshipit-source-id: a32c85b43432f7ee8391c7dff625fdbe0eacf312