test: Add IntermediateTypeTransforms for supporting intermediate types in PrestoQueryRunner (#12746)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/12746

Presto supports a number of what I call "intermediate types" that are supported in the query
engine but not in the file formats.  This means we can't currently use fuzzers with the
PrestoQueryRunner with these types because there's no way to ensure we don't generate
plans with these types as inputs (which have to be written to a DWRF file), and if we did
exclude them, that would prevent us from e.g. testing them in keys/aggregates in the
AggregationFuzzer which doesn't do any projections.

To solve this, I add IntermediateTypeTransforms which provides functions transform and
projectionExpr. Transform converts values of an "intermediate type" to values of a type that
is supported in the file format. projectionExpr generates an expression tree that can convert
the types output by transform back into the "intermediate type".

In a follow up diff I integrate this into the fuzzers that support PrestoQueryRunner so that
they can begin fuzzing "intermediate types".

In this diff I provide implementations for Array, Map, and Row in cases where they contain an
"intermediate type", as well as an implementation for TimestampWithTimeZone as the
inaugural "intermediate type".

Reviewed By: kagamiori

Differential Revision: D71585451

fbshipit-source-id: c8ce1cf4572d8e6e48e48b2745ea4b00fc879a57