Support execution by operator for table functions Adds support for table functions that have table inputs. Execution by operator for table functions without sources is not yet implemented.
If the input is empty, and it does not have the
KEEP WHEN EMPTY property, the function is not executed,
and empty result is returned.

If the input is empty, and it has the KEEP WHEN EMPTY
property, the function is executed upon the EmptyTableFunctionPartition
It is ensured that the function with empty input is executed once per node,
and that there is single node distribution in case of the KEEP WHEN EMPTY
property.

Changes adapted from trino/PR#15575
Original commit: c2e838e7647b452e490a012d9af2aca2f4ac8e3a
Author: kasiafi

Co-authored-by: Xin Zhang <desertsxin@gmail.com>
Co-authored-by: kasiafi <30203062+kasiafi@users.noreply.github.com>