fix: ExpressionFuzzer never selects function signatures that return a type variable (#12933)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/12933

I was trying to run the expression fuzzer on the newly added array_min_by/array_max_by
functions and discovered it would never select either of these functions when generating an
expression.

I discovered that when we look up function templates as candidates, we use
`typeToExpressionList_[kTypeParameterName]`
Likewise we use kTypeParameterName when looking up a signature in
expressionToTemplatedSignature_ with chooseRandomSignatureTemplate.

However, when we add entries to these two maps, we first sanitize the type name, which just
converts it to lower case. kTypeParameterName is "T" which will clearly never match an all
lower case string. So in practice we never recognize functions that return a type variable as a
viable candidate.

This diff update kTypeParameterName to sanitize the string we choose for it, to ensure it's
consistent with the keys in the maps we use it for lookups in.

I also noticed that when we iterate over the list of signatureTemplates_, we check if the return
type is a type variable, but again we sanitize it before doing this check.  Since simple
functions use strings like __user_T1 for type variable names, and it's common practice to use
capital letters like T and U in Vector functions, this likewise never matches.

I updated this logic to do the check against the unsanitized type name as well.

These bugs were preventing ExpressionFuzzer from ever covering array_min_by,
array_max_by, subscript, find_first at all, among other functions that are only being partially
covered.

Reviewed By: kagamiori

Differential Revision: D72474230

fbshipit-source-id: 717fc403ce6deb5c842ca58a14c1af3bae3f3415