Add configuration to specify metastore catalog name
### Description:
This commit introduces a new configuration to allow users to specify a catalog name for the metastore in Presto. By using the `hive.metastore.catalog.name` property, users can manage and connect to multiple metastores with the same catalog name. This enhancement enables a more flexible and configurable approach to handling catalogs and schemas across different metastores and storage backends, especially in complex, multi-region, and multi-tenant environments.