Disable modernizer plugin on each module
This could be done in a single file, but by splitting this up it
will allows contributors and maintainers to incrementally re-
enable the modernizer on each module since it will cause lots of
code change. It will also introduce lots of opportunities for new
contributors to get involved with Presto.