Add operator < to ArrowColumnHandle protocol
ArrowColumnHandle was missing the operator<() in the protocol definition.
This was causing queries that require ordering of columns to fail. A special
definition of ArrowColumnHandle is required in the protocol to define
the operator to order columns by the columnName.

Add e2e test for ColumnHandle ordering

Add a Java unit test to test the native ArrowFlight connector will
support queries that require ordering of column names.