[native] Fix blank HTTP host header in requests
Because the request's destination address wasn't set inside
of the RequestBuilder object, the Host was populated with
a blank value. On older HTTP versions this is permissible, but
on newer ones (such as Jetty 12), this is a hard error and results
in the request being rejected with a 400 error.

This change properly sets the request destination address so
that ensureHostHeader sets the correct value.