fix: Inverse_poisson_cdf doesn't handle case where results exceed bounds of int32_t (#13048)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/13048

inverse_poisson_cdf doesn't handle the case where results exceed the bounds of int32_t, so
it throws an exception (at least with sanitizers enabled) when converting the result to an
int32_t.

Presto Java handles this by bounding the result to Integer.MAX_VALUE. The change applies
the same restriction to Velox's implementation.

Reviewed By: skyelves

Differential Revision: D73143995

fbshipit-source-id: 9bb2224ae37ccae8e5bf4c8a6f3872882b51469c