fix: Make precision recall functions always use size() for termination condition (#13056)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/13056

Precision recall functions output a list of doubles based on the used size of the bucket. The original implementation copied Java's termination condition for outputting the aggregated values here: https://github.com/prestodb/presto/blob/47a708022aba003ebfda483ea3c940231fdbd282/presto-main-base/src/main/java/com/facebook/presto/operator/aggregation/PrecisionRecallAggregation.java#L205

which translates to index < bucketCount && totalTrueWeight > runningTrueWeight. However, this is equivalently used to what I introduced as `maxUsedIndex_`, which is the last index in the bucketCount. This accomplishes the condition in the Java implementation, and avoids an O(bucketCount) computation.

This optimization actually exposed a bug in Java because of precision loss during decimal adding, causing more terms in Java to returned.

Made up fake example:
- bucket_count = 5
- true_weights = [0, 0.01, 0.02, 0, 0]
- total_true_weigh_sum = 0.03

It won't happen in this example, but you can imagine when we add the 3rd index to the running sum, we can get something like 0.0299999. As a result of this in-precision adding, we won't stop at 3 but will return bucket_count (5). This is what happens in Java.

See https://github.com/prestodb/presto/issues/24936 for more detail

In this diff, alway use `size()`, the max index used, which is more correct instead of using Java's implementation.

Reviewed By: kletkavrubashku

Differential Revision: D73084745

fbshipit-source-id: 3a6dc0dcddc34e84043a1dec743b8c8960f704af