fix: Fix corner case in processConstantFilterResults() (#13074)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/13074

processFilterResults() from OperatorUtils is being used in HiveDataSource to filter data coming from the SplitReader.
We call is from HiveDataSource::evaluateRemainingFilter().
Normally, it returns the number of rows remaining after running the remaining filter and also it populates indices in the filterEvalCtx_.selectedIndices.
These indices are then used to wrap child vectors of the output row vector into Dictionaries in case some rows were filtered out.

There is a corner case when HiveDataSource applies bucket conversion (which can filter some rows and then it applies the remainig filter).
In this corner case the code path does not generate the new selected indices and HiveDataSource wraps vectors using the selected indices from the previous batch, which can result in correctness issues or crashes.

We hit a corner case when all these conditions are met:
- We have bucket conversion (I don't know what it is exactly, but looks like an extra filtering per thread based on the bucketed column when only single bucket is allowed, could be to improve following aggegation performance).
- Bucket conversion did not filter out all rows (some rows passed it).
- We have remaining filter.
- For all rows in a batch which passed bucket conversion filter we have remaining filter evaluates to true, thus causing the filterResult_ encoding to be constant.

To fix the issue we adding a check in processConstantFilterResults() to see if we are reducing the selection compared to the filterResult.
If we do, then we need to fill up the selected indices.

Reviewed By: kewang1024

Differential Revision: D73288408

fbshipit-source-id: 62c3665747afd7b41a0634b7c2c2dc3f10da734e