fix: Change velox buffer copy and serialization interfaces to take size in i64 instead of i32 (#13083)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/13083

Many internal queries crash at places like crc computation, or serialization because we have an restriction on the size of the buffer since the input to many of these functions are int32_t. Let's change this to be int64_t and let the application specific logic take care of the large buffer if need be.

The components/functions changed are:
- A few simd utils (memcopies)
- newView API in byteStream
- process_bytes in the crc computation
- byteRange struct
- Functions in ByteStream to internally use int64_t
- StreamArena's newRange and allocate

For example, Presto queries that would crash at crc computation or serialization time would now throw memory exceeded or spill error similar to Java.

Reviewed By: xiaoxmeng, kewang1024

Differential Revision: D73310051

fbshipit-source-id: 1dc2593be81ca74b07601fb51b76b0a81cfb72d3