Fix compilation for presto-testing-mysql with io.airlift.units
presto-testing-mysql still depends on io.airlift:units instead of
com.facebook.airlift:units. This causes issues for the tests which
set MySqlOptions since the builder type only accepts units. Rather than
include the io.airlift:units dependency, this commit removes the use of
the builder options which accept a Duration type in order to allow the
project to compile.

This commit can be reverted if we release a new version of
presto-testing-mysql with the new com.facebook.airlift:units dependency.